Rename Solsock to Solsocket and make it type sockproto
This constant is used as a protocol level in getsockopt/setsockopt as well as
msghdr. Valid values for level include the same constants used as the protocol
argument of socket (IPPROTO_*), so it needs to have the same type so the
constants can be used in both functions.

The rename is for consistency with the C API.