let bychunk() read the last element; guard for negative sz
I mentioned on IRC that bychunk() was skipping the last element of
a slice, here's a fix for that, as well as some redundant tests and
some probably unneeded input validation. I don't have any strong
feelings about protecting someone from shooting themselves in the
foot with bychunk([1][:], -100), but it's easier to remove in review
than to add later.