Subnormalize special-case floats in fltXYassem.
If the exponent is the lowest possible representable, then the
returned float will actually be subnormal. In this case, we do not
have the implied leading 1, so we need to downshift the significand
so that it doesn't get lost. We end up losing the least significant
bit of the significand, but that's unavoidable.