Implement most of expm1. Errors in huge numbers.
[Tan92] finishes off calculation of expm1(f) by range expansion,
which involves multiplying a number by 2^M, where M grows roughly
linearly with f. If f is very large (say, 0x0x42b17200 in
single-precision), M may grow to be as larger than the maximal
exponent (in this case, M = 128).

If 2^M * P is calculated in the normal way, the intermediate value
2^M is not representable, causing mangled output. However, the
result may still be computed: the value P is of magnitude < 1.
Therefore, a scalb function will be necessary to perform this
multiplication without creating unrepresentable intermediate values={