Use fma instead of rote multiplication in powr's final calculation.
Since the exp() we use is sometimes wrong by 1 ulp, the final
multiplication has a possibility of increasing that to 2 ulps. Using fma
instead makes such a thing much less likely.

This makes some of our off-by-0s off-by-1s, and vice versa, but on the
whole it is a definite improvement.